+++ app/Http/Controllers/Frontend/FrontendPageController.php
@@ -185,6 +185,23 @@ class FrontendPageController extends Controller
         }
         return redirect()->back()->with('success', 'บันทึกสำเร็จ !');
     }
+    public function updatecontackbackPage(Request $request)
+    {
+        $postId = $request->id;
+        $currentValue = $request->currentValue;
+        $newValue = $currentValue == 'create' ? 'contact' : 'create';
+        $newValuefornotice = $currentValue == 'create' ? 'read' : 'create';
+        contacts_backModel::where('id', $postId)->update([
+            'status' => $newValue,
+        ]);
+        noticeModel::where('resource_id', $postId)
+        ->where('resource', 'contacts_back')
+        ->update([
+            'status' => $newValuefornotice,
+        ]);
+
+        return response()->json(['status' => 'success', 'newValue' => $newValue]);
+    }
     public function updatereservePage(Request $request)
     {
         $postId = $request->id;
@@ -846,65 +863,46 @@ class FrontendPageController extends Controller
         }
 
         // $contacts_back = contacts_backModel::where("customer_id", $customer_id)->get();
-        $contacts_back = DB::table('contacts_back')
-        ->select('contacts_back.*', 'cars.*', 'brands.title as brand_title', 'models.model as model_name', 'customer.*')
-        ->where('contacts_back.customer_id', '=', $customer_id)
-        ->join('cars', 'contacts_back.cars_id', '=', 'cars.id')
-        ->join('brands', 'cars.brand_id', '=', 'brands.id')
-        ->join('models', 'cars.model_id', '=', 'models.id')
-        ->join('customer', 'contacts_back.customer_id', '=', 'customer.id')
-        ->get();
+        // $contacts_back = DB::table('contacts_back')
+        // ->select('contacts_back.*', 'cars.*', 'brands.title as brand_title', 'models.model as model_name', 'customer.*')
+        // ->where('contacts_back.customer_id', '=', $customer_id)
+        // ->join('cars', 'contacts_back.cars_id', '=', 'cars.id')
+        // ->join('brands', 'cars.brand_id', '=', 'brands.id')
+        // ->join('models', 'cars.model_id', '=', 'models.id')
+        // ->join('customer', 'contacts_back.customer_id', '=', 'customer.id')
+        // ->get();
 
-        // $query_contact_back = DB::table('contacts_back')
-        //     ->select(
-        //         'contacts_back.*', 
-        //         'cars.id as car_id', // Include the id of cars
-        //         'cars.*', 
-        //         'brands.title as brand_title', 
-        //         'models.model as model_name', 
-        //         'customer.*'
-        //     )
-        //     ->where('contacts_back.customer_id', '=', $customer_id)
-        //     ->join('cars', 'contacts_back.cars_id', '=', 'cars.id')
-        //     ->join('brands', 'cars.brand_id', '=', 'brands.id')
-        //     ->join('models', 'cars.model_id', '=', 'models.id')
-        //     ->join('customer', 'contacts_back.customer_id', '=', 'customer.id')
-        //     ->get();
 
         $query_contact_back = DB::table('contacts_back')
-            ->select('contacts_back.*', 'cars.*', 'brands.title as brand_title', 'models.model as model_name', 'customer.*')
+            ->select('contacts_back.id as contact_id', 'contacts_back.status as contact_status', 'contacts_back.*', 'cars.id', 'cars.status', 'cars.customer_id', 'cars.user_id', 
+            'cars.type', 'cars.brand_id', 'cars.model_id', 'cars.modelyear', 'brands.title as brand_title', 
+            'models.model as model_name', 'customer.*', 'contacts_back.created_at')
             ->join('cars', 'contacts_back.cars_id', '=', 'cars.id')
             ->join('brands', 'cars.brand_id', '=', 'brands.id')
             ->join('models', 'cars.model_id', '=', 'models.id')
             ->join('customer', 'cars.customer_id', '=', 'customer.id')
             ->where('cars.customer_id', '=', $customer_id)
-            ->get();
+            ->orderBy('contacts_back.id', 'desc')
+            ->paginate(24);
 
 
-
-
-
-
-
-
-
-        $mycontacts = contacts_backModel::select(
-            'contacts_back.*', // Select all fields from contacts_back
-            'customer.id as customer_id',
-            'cars.*', // Select all fields from cars
-            'brands.title as brand',
-            'models.model as model', // Replace models.title with models.model
-            'generations.generations as generation', // Replace generations.title with generations.generations
-            'sub_models.sub_models as sub_model' // Replace sub_models.title with sub_models.sub_models
-        )
-        ->join('cars', 'contacts_back.cars_id', '=', 'cars.id')
-        ->join('customer', 'contacts_back.customer_id', '=', 'customer.id')
-        ->join('brands', 'cars.brand_id', '=', 'brands.id')
-        ->join('models', 'cars.model_id', '=', 'models.id')
-        ->join('generations', 'cars.generations_id', '=', 'generations.id')
-        ->join('sub_models', 'cars.sub_models_id', '=', 'sub_models.id')
-        ->orderBy('contacts_back.created_at', 'desc')
-        ->get();
+        // $mycontacts = contacts_backModel::select(
+        //     'contacts_back.*', // Select all fields from contacts_back
+        //     'customer.id as customer_id',
+        //     'cars.*', // Select all fields from cars
+        //     'brands.title as brand',
+        //     'models.model as model', // Replace models.title with models.model
+        //     'generations.generations as generation', // Replace generations.title with generations.generations
+        //     'sub_models.sub_models as sub_model' // Replace sub_models.title with sub_models.sub_models
+        // )
+        // ->join('cars', 'contacts_back.cars_id', '=', 'cars.id')
+        // ->join('customer', 'contacts_back.customer_id', '=', 'customer.id')
+        // ->join('brands', 'cars.brand_id', '=', 'brands.id')
+        // ->join('models', 'cars.model_id', '=', 'models.id')
+        // ->join('generations', 'cars.generations_id', '=', 'generations.id')
+        // ->join('sub_models', 'cars.sub_models_id', '=', 'sub_models.id')
+        // ->orderBy('contacts_back.created_at', 'desc')
+        // ->get();
 
 
         
@@ -921,8 +919,8 @@ class FrontendPageController extends Controller
             'customer_id' => $customer_id,
             'mycars' => $mycars,
             'carfromstatus' => $carfromstatus,
-            'contacts_back' => $contacts_back,
-            'mycontacts' => $mycontacts,
+            // 'contacts_back' => $contacts_back,
+            // 'mycontacts' => $mycontacts,
             'carstatus' => "approved",
             'brandsearch' => $qrybrandsearch,
             'query_contact_back' => $query_contact_back,
@@ -1183,9 +1181,11 @@ class FrontendPageController extends Controller
         }
         if($request->hasFile('map')){
 
-            $oldPath = public_path($Customer->map);
-            if(File::exists($oldPath)){
-                File::delete($oldPath);
+            if(isset($Customer->map)){
+                $oldPath = public_path($Customer->map);
+                if(File::exists($oldPath)){
+                    File::delete($oldPath);
+                }
             }
 
             $file = $request->file('map');
